confirm delete project message

+++ ppm-react-app/src/actions/projectActions.js
@@ -38,9 +38,11 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = id => async dispatch => {
-	await axios.delete(`/api/project/${id}`);
-	dispatch({
-		type: DELETE_PROJECT,
-		payload: id,
-	});
+	if (window.confirm('Delete Project?')) {
+		await axios.delete(`/api/project/${id}`);
+		dispatch({
+			type: DELETE_PROJECT,
+			payload: id,
+		});
+	}
 };
